Working on LU Decomposition.
Added comments to clarify parts of the code. Still clarifying the back
substituation function.

Added order vector to swap rows. Had to add this everywhere a row is referenced.

May create two different functions. One to calculate the L and U matricies and
one to calculate the back substitution. This will work well when the back
substitution step needs to be done but not the whole LU decomposition. Will also
need a function to multiply matricies; not sure if it will write this one or let
numpy handle it.

Update the indexing so hopefully the first column and row will be 0 instead of 1
like the algorithem I am following has done.