Finished adding some docs to the back_sub. Added multiple_rhs func.
Added clarifications to the back_sub. Still not a clear as I would like it to
be.

Started thinking about how to implement the multiple right hand side solver. I
may want to leave this alone for now. Develop what I need now and come back to
this later. I will probably leave it in the code as a reminder.

Need to see if it runs and set up some testing. Need to leave the multiple rhs
solver alone for now.