We have output! It's not right. Still got some work to do.
Something is broken and it is going to take some more stepping through the pdb
debugger and probably some hand calcs to get it straightened out.

I took a look at the scipy linalg lu decomposition package and it seems like it
does something similar to what I am proposing. Having a function that do the
decomposition of A. Then separate function to apply equivalent operations the
rhs and another function to back substitute to solve for x. There is also the
option which I was considering to be able to apply the operations to many rhs's
and solve for many x's without decompositing the matrix again.

Gotta get something working first thought. :)