Solves basic example. RHS operations still coupled to LU decomp.
Got the back substitution working. Logic was correct just getting the right
ranges for the for-loops figured out. Had to walk through it with pdb debugger.

The RHS operations are still part of the main lu_decomp function. This operation
needs to be pulled out. Even if it slows the whole solution down a bit it will
make a lot of sense to decouple the LU decomposition of the A matrix from the
application of those operations to the b vector.

The solution, finding the x vector, can be found with back substitution. It can
also be found in another way that I am not as familiar with but it has to do
with the L and U matricies. I think two functions could be set up, one that
solves for x directly and another that is provided the L and U matricies. The
first is found with normal back substitution. The second is found in the
alternate way.

Need to do research on how to solve for additional RHSs.