Update naming convention, docs, clean up functions.
Came up with a new naming convention for functions. Taking after the likes
of BLAS and LINPACK but not quite the same. Realized the 'lu_decomp' is not
a great name because it doesn't give insight to the algorithm used to do the
decomposition.

Updated other names to be more generic.

Cleaned up doc strings. Spyder has a nice utility to create the doc formatting
automatically.

Cleaned up functions. Removed unnecessary bits and pieces like the 'nmo' which
I think was a hold over from a VBA routine I wrote or something.