Multiple updates see full commit message.
(Really need to do a better job breaking up commits.)

First, change gega to gedo. Apparently, it's not Gaussian elimination.
When the steps are recorded and it is not appended with a b-vector r.h.s.
it is a Doolittle reduction.

Also, made the pivot optional in the gedo function.

Second, started working a Crout reduction function. It is not yet
complete. Basic functionality works, no pivoting and the forward
and back substitution don't work right yet.