Further updates w.r.t pivoting error.
I wanted to understand this bug better so I reverted the
code and stepped through it until I understood in detail
why it was happening.

I decided to rename variables to better clarify their intention
and the error becomes even more obvious when the variables are
renamed. I also added some additional comments to provide further
clarification because it is a bit confusing.