Change gecr function exception to warning.
Discovered for the Doolittle reduction that a warning should be used
when a near-zero value is found on the diagonal instead of an
exception. It is up to the user to handle this situation. The math is
possible so no exception should be raised. However, knowing that in
the vast majority of use cases users want the matrix to be stable and
well-conditioned, I think it is a good idea to raise a warning to
alert when this is not the case.

Added this to the Crout reduction algorithm