Major additions and updates to linalg.py.
This definately should have been multiple commits.

Apologies. (Probably mostly to my future self)

1. Added a try block in gedo() function to catch an error and provide
a useful error message in the event of division by zero with a zero
on the diagonal.

2. Add a function to do a full inverse. This was initially added to the
linalg_exp.py. I didn't think it would be necessary to add the full
inverse to the non-experiemental module but I did anyway in case
someone wants to use it. Although I can't think of a reason why.

3. Added formatting for tri-diagonal and crout reduction functionality.

4. Slight change to the docstrings for the ddm() function.

5. Added two functions for the solution of a system of nonlinear equations.
The first function newton_soe() is an implementation of Newton's method
for multiple equations updating the Jacobian matrix on each iteration. The
second is a modified Newton's method for multiple equations where the Jacobian
matrix is updated after every nth iteration.

Ok, not quite as bad as I thought but really should have been doing a better
job with the intermediate commits.