Add linalg_exp.py module.
linalg_exp.py is an experimental version of linalg.py which is the production
verison.

I created the linalg_exp.py module so I can more easily modify functionally as
needed for experimentation and textbook problems. To allow for more verbose output,
specification of precision to understand how it influences accuracy of results.

It is basically a copy I can modify to my hearts content without having to worry
about how it will affect the production module. There may be a better way to this
with a git branch or something clever but this is working for my purposes.