Converted all textbook solution .py files to .ipynb. Using .ipynb now.
This commit includes:
1. The migration of .py to .ipynb files.

2. Additional worked textbook problems in the .ipynb format that were done after
the decision to migrate to .ipynb.

With the inline graphing and other nice features of the Jupyter notebook platform
I decided to migrate all the textbook solutions from .py to .ipynb.

I will be moving forward using Juypter notebooks for the textbook problems. I will,
however, continue to develop the core modules in .py and reference those into
the notebooks for actual use.

I plan to try to commit individual textbook problems from now on.
As with a previous commit I recognize this is not ideal and a rather large
amount of changes to commit all at once.