Updated formal parameters, added polynomial degree
Updated the lag_poly function to accept the polynomial degree
as an argument. I guess the is somewhat inherent in the numbers
of points provided but I think it makes it more explicit to the
user. It also allows for a check to make sure the user supplies
a sufficient number of points to model the polynomial degree
indicated.