Completed Ex 3.4 as part of ch3 p4.
Walked through this example to understand how to do
a significant portion of ch3 p4.

Understanding the values of the different derivatives
calculated at minimum and maximum values and how the
the upper and lower error bounds are calculated took
some reverse engineering of the table provided in this
example.