Added two direct matrix solution functions.
Added a direct solution routine for n x n matricies using the other
available functions in the package. Effectively, instead of returning
the LU matrix and corresponding order vector which is useful for
multiple solutions, this function just solves once and returns the
solution. All prior functionality was left in place. This new function
makes it easier on the user to make one function call for one solution.

A similar function was written for tridiagonals.

Written in response to work I was doing with cubic spline function
routines where only one solution is needed.