Added naive, basic matrix routines.
I realized I had not implemented basic matrix routines such as
matrix tranpose, vector-vector, matrix-vector and matrix-matrix
multiplication.

I have added naive implementations of each with the intention
of returning later to explore more sophisticated and efficient
algorithms such as loop nest optimization to optimize the cache
memory and divide-and-conquer which is a recursive algorithm and
cache-oblivious.