Updated and added clarifications about returned solution order.
Updated the returned solutions to correspond with the original order of
equations in selected functions, specifically the doqsv. Updated other
functions with documentation commentary to explain that the returned
solution order does not correspond to the original equation order and
how to easily return the solution order to the original equation order.